refactor:#5 모호한 파일명 수정, 불필요하게 분리했던 파일 합체
