+++ src/constants/naverMap.js
@@ -1,11 +1,3 @@
-export const supabaseItems = {
-  PLANS: 'plans',
-  DETAILPLANS: 'detail_plans',
-  PLANDATE: 'plan_date',
-  PLANTIME: 'plan_time',
-  ID: 'id'
-};
-
 export const mapCreateFile = {
   ZOOM: 14
 };
