+++ src/constants/supabaseConstants.js
@@ -1,0 +1,7 @@
+export const supabaseItems = {
+  PLANS: 'plans',
+  DETAILPLANS: 'detail_plans',
+  PLANDATE: 'plan_date',
+  PLANTIME: 'plan_time',
+  ID: 'id'
+};
